backend/mac_requirements.txt
aniso8601==7.0.0
Django==2.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
graphene==2.1.8
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==8.1.2
promise==2.3
pytz==2021.1
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
